requirements.txt
# Core dependencies for Crawl4AI RAG MCP Server
sentence-transformers==3.2.1
sqlite-vec==0.1.6
numpy==2.2.1
requests==2.32.5
gliner==0.2.15
pyyaml==6.0.2

# REST API dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1

# Optional dependencies for production
gunicorn==23.0.0
prometheus-client==0.21.1

# Development dependencies
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
flake8==7.1.1